ref: extract SDK-based profiling tests to separate test class in Profiling test suite
