+++ Tests/SentryTests/SentrySDKTests.swift
@@ -947,368 +947,6 @@ class SentrySDKTests: XCTestCase {
 #endif // SENTRY_HAS_UIKIT
 }
 
-#if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
-extension SentrySDKTests {
-    func testStartingContinuousProfilerWithSampleRateZero() throws {
-        givenSdkWithHub()
-
-        fixture.options.profilesSampleRate = 0
-        XCTAssertEqual(try XCTUnwrap(fixture.options.profilesSampleRate).doubleValue, 0)
-
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        SentrySDK.startProfiler()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingContinuousProfilerWithSampleRateNil() throws {
-        try withMockTimerFactory { timerFactory in
-            givenSdkWithHub()
-
-            // nil is the default initial value for profilesSampleRate, so we don't have to explicitly set it on the fixture
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-            SentrySDK.startProfiler()
-            XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-            // clean up
-            try stopProfiler(timerFactory: timerFactory)
-        }
-    }
-
-    func testNotStartingContinuousProfilerWithSampleRateBlock() throws {
-        givenSdkWithHub()
-
-        fixture.options.profilesSampler = { _ in 0 }
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        SentrySDK.startProfiler()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testNotStartingContinuousProfilerWithSampleRateNonZero() throws {
-        givenSdkWithHub()
-
-        fixture.options.profilesSampleRate = 1
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        SentrySDK.startProfiler()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingAndStoppingContinuousProfiler() throws {
-        let timerFactory = TestSentryNSTimerFactory(currentDateProvider: fixture.currentDate)
-        let originalTimerFactory = SentryDependencyContainer.sharedInstance().timerFactory
-        SentryDependencyContainer.sharedInstance().timerFactory = timerFactory
-
-        givenSdkWithHub()
-        SentrySDK.startProfiler()
-        XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-        try stopProfiler(timerFactory: timerFactory)
-
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        SentryDependencyContainer.sharedInstance().timerFactory = originalTimerFactory
-    }
-
-    func testStartingContinuousProfilerBeforeStartingSDK() {
-        SentrySDK.startProfiler()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingContinuousProfilerAfterStoppingSDK() {
-        givenSdkWithHub()
-        SentrySDK.close()
-        SentrySDK.startProfiler()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testManuallyStartingAndStoppingContinuousProfilerV2Sampled() throws {
-        // arrange
-        try withMockTimerFactory { mockTimerFactory in
-            try withMockRandom(value: 0.5) {
-                fixture.options.profiling.sessionSampleRate = 1
-                givenSdkWithHub()
-                sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-                // act
-                SentrySDK.startProfileSession()
-
-                // assert
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // act
-                try stopProfilerV2(timerFactory: mockTimerFactory)
-
-                // assert
-                XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-            }
-        }
-    }
-
-    func testManuallyStartingAndStoppingContinuousProfilerV2NotSampled() throws {
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.sessionSampleRate = 0
-            givenSdkWithHub()
-            SentrySDK.startProfileSession()
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        }
-    }
-
-    func testStartingContinuousProfilerV2BeforeStartingSDKDoesNotStartProfiler() {
-        SentrySDK.startProfileSession()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingContinuousProfilerV2AfterStoppingSDKDoesNotStartProfiler() {
-        givenSdkWithHub()
-        SentrySDK.close()
-        SentrySDK.startProfileSession()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingContinuousProfilerV2WithoutContinuousProfilingEnabledDoesNotStartProfiler() {
-        fixture.options.profilesSampleRate = 1
-        givenSdkWithHub()
-        SentrySDK.startProfileSession()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingContinuousProfilerV2WithTraceLifeCycleDoesNotStartProfiler() {
-        fixture.options.profiling.lifecycle = .trace
-        givenSdkWithHub()
-        SentrySDK.startProfileSession()
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testContinuousProfilerV2TraceLifecycleZeroSampleRateDoesNotStartProfiler() throws {
-        // arrange
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.lifecycle = .trace
-            fixture.options.profiling.sessionSampleRate = 0
-            fixture.options.tracesSampleRate = 1
-            sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-            givenSdkWithHub()
-            fixture.currentDate.advance(by: 1)
-            
-            // act
-            let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-            
-            // assert
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-            
-            // clean up
-            trace.finish()
-        }
-    }
-
-    func testStoppingContinuousProfilerV2WithTraceLifeCycleDoesNotStopProfiler() throws {
-        // arrange
-        try withMockTimerFactory { mockTimerFactory in
-            try withMockRandom(value: 0.5) {
-                fixture.options.profiling.lifecycle = .trace
-                fixture.options.profiling.sessionSampleRate = 1
-                fixture.options.tracesSampleRate = 1
-                sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-                givenSdkWithHub()
-                fixture.currentDate.advance(by: 1)
-                let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // act - using manual stop method is a no-op in trace profile lifecycle mode
-                try stopProfilerV2(timerFactory: mockTimerFactory)
-
-                // assert
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // arrange
-                fixture.currentDate.advance(by: 1)
-
-                // act - the current profile chunk will automatically finish
-                trace.finish()
-
-                // assert - profile chunk must complete
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // arrange - simulate chunk completion
-                fixture.currentDate.advance(by: 60)
-                try mockTimerFactory.check()
-
-                // assert - profiler is stopped
-                XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-            }
-        }
-    }
-
-    func testStartingTransactionWithoutTraceLifecycleDoesNotStartContinuousProfilerV2() {
-        fixture.options.tracesSampleRate = 1
-        givenSdkWithHub()
-        SentrySDK.startTransaction(name: "test", operation: "test")
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-    }
-
-    func testStartingTransactionWithTraceProfilingLifecycleWithTracingDisabledDoesNotStartContinuousProfilerV2() {
-        fixture.options.profiling.lifecycle = .trace
-        givenSdkWithHub()
-
-        // hold a reference to the tracer so it doesn't dealloc, which tries to clean up any existing profilers
-        let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-        XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        trace.finish()
-    }
-
-    func testContinuousProfilerV2ManualLifecycleStartWithSampleSessionDecisionYes() throws {
-        // arrange
-        try withMockTimerFactory { timerFactory in
-            try withMockRandom(value: 0.5) {
-                fixture.options.profiling.sessionSampleRate = 1
-                givenSdkWithHub()
-                sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-                // act
-                SentrySDK.startProfileSession()
-
-                // assert
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // clean up
-                try stopProfilerV2(timerFactory: timerFactory)
-            }
-        }
-    }
-
-    func testContinuousProfilerV2ManualLifecycleStartWithSampleSessionDecisionNo() throws {
-        // arrange
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.sessionSampleRate = 0
-            givenSdkWithHub()
-            sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-            // act
-            SentrySDK.startProfileSession()
-
-            // assert
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-        }
-    }
-
-    func testContinuousProfileV2TraceLifecycleTracesSampleDecisionYesSessionSampleDecisionNo() throws {
-        // arrange
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.sessionSampleRate = 0
-            fixture.options.profiling.lifecycle = .trace
-            fixture.options.tracesSampleRate = 1
-            givenSdkWithHub()
-            sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-            // act
-            let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-
-            // assert
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-
-            // clean up
-            trace.finish()
-        }
-    }
-
-    func testContinuousProfileV2TraceLifecycleTracesSampleDecisionNoSessionSampleDecisionNo() throws {
-        // arrange
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.sessionSampleRate = 0
-            fixture.options.profiling.lifecycle = .trace
-            fixture.options.tracesSampleRate = 0
-            givenSdkWithHub()
-            sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-            // act
-            let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-
-            // assert
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-
-            // clean up
-            trace.finish()
-        }
-    }
-
-    func testContinuousProfileV2TraceLifecycleTracesSampleDecisionYesSessionSampleDecisionYes() throws {
-        // arrange
-        try withMockTimerFactory { mockTimerFactory in
-            try withMockRandom(value: 0.5) {
-                fixture.options.profiling.sessionSampleRate = 1
-                fixture.options.profiling.lifecycle = .trace
-                fixture.options.tracesSampleRate = 1
-                givenSdkWithHub()
-                sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-                // act
-                let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-
-                // assert
-                XCTAssert(SentryContinuousProfiler.isCurrentlyProfiling())
-
-                // clean up
-                trace.finish()
-                fixture.currentDate.advance(by: 60)
-                try mockTimerFactory.check()
-                XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-            }
-        }
-    }
-
-    func testContinuousProfileV2TraceLifecycleTracesSampleDecisionNoSessionSampleDecisionYes() throws {
-        // arrange
-        try withMockRandom(value: 0.5) {
-            fixture.options.profiling.sessionSampleRate = 1
-            fixture.options.profiling.lifecycle = .trace
-            fixture.options.tracesSampleRate = 0
-            givenSdkWithHub()
-            sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
-
-            // act
-            let trace = SentrySDK.startTransaction(name: "test", operation: "test")
-
-            // assert
-            XCTAssertFalse(SentryContinuousProfiler.isCurrentlyProfiling())
-
-            // clean up
-            trace.finish()
-        }
-    }
-}
-
-private extension SentrySDKTests {
-    func withMockTimerFactory(block: (TestSentryNSTimerFactory) throws -> Void) throws {
-        let timerFactory = TestSentryNSTimerFactory(currentDateProvider: fixture.currentDate)
-        let originalTimerFactory = SentryDependencyContainer.sharedInstance().timerFactory
-        SentryDependencyContainer.sharedInstance().timerFactory = timerFactory
-
-        try block(timerFactory)
-
-        SentryDependencyContainer.sharedInstance().timerFactory = originalTimerFactory
-    }
-
-    func withMockRandom(value: Double, block: () throws -> Void) throws {
-        let random = TestRandom(value: value)
-        let originalRandom = SentryDependencyContainer.sharedInstance().random
-        SentryDependencyContainer.sharedInstance().random = random
-        
-        try block()
-
-        SentryDependencyContainer.sharedInstance().random = originalRandom
-    }
-
-    func stopProfiler(timerFactory: TestSentryNSTimerFactory) throws {
-        SentrySDK.stopProfiler()
-        fixture.currentDate.advance(by: 60)
-        try timerFactory.check()
-    }
-
-    func stopProfilerV2(timerFactory: TestSentryNSTimerFactory) throws {
-        SentrySDK.stopProfileSession()
-        fixture.currentDate.advance(by: 60)
-        try timerFactory.check()
-    }
-}
-#endif // os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
-
 private extension SentrySDKTests {
     func givenSdkWithHub() {
         SentrySDK.setCurrentHub(fixture.hub)
